refactor: remove server-side acl permission system and shared agent ui
remove the ACL layer (public/email permissions) from the access control model,
simplifying to owner + org member only. This is part B of the ACL cleanup (#4871).

- create simplified canAccessCompose() in compose-access.ts (owner + org only)
- delete permission-service.ts, permission-sync.ts, and permissions CRUD route
- remove authorizeCompose() from run-service.ts (access checked at API layer)
- remove addPermission/removePermission calls from Slack/Telegram integrations
- remove getEmailSharedAgents and shared agent logic from list/get-user-agents
- remove isOwner from API contract (composeListItemSchema) and all frontend refs
- remove "Shared" badge, non-owner read-only mode, and org/name splitting in UI
- update tests to remove ACL-specific test cases

Closes #4875
