+++ turbo/apps/platform/src/mocks/handlers/api-agents.ts
@@ -18,7 +18,6 @@ export const apiAgentsHandlers = [
           displayName: null,
           headVersionId: "version_1",
           updatedAt: "2024-01-01T00:00:00Z",
-          isOwner: true,
         },
       ],
     });
