+++ turbo/apps/platform/src/views/integrations-page/github-settings-page.tsx
@@ -139,36 +139,8 @@ export function GitHubSettingsPage() {
   const openConfirm = useSet(openGithubDisconnectDialog$);
   const closeConfirm = useSet(closeGithubDisconnectDialog$);
 
-  // Construct qualified agent name that matches the format used in agentsList$
-  // (shared agents use "org/name", owned agents use just "name").
-  const qualifiedAgentName = (() => {
-    if (!data?.agent) {
-      return undefined;
-    }
-    const fullName = `${data.agent.orgSlug}/${data.agent.name}`;
-    // If the qualified name exists in agents list, use it (shared agent)
-    if (agents.some((a) => a.name === fullName)) {
-      return fullName;
-    }
-    // Otherwise use bare name (owned agent)
-    return data.agent.name;
-  })();
-
-  // Ensure the current agent appears in the dropdown even if
-  // it isn't in the user's own agents list (e.g. shared by another user).
-  const agentOptions = (() => {
-    if (!qualifiedAgentName) {
-      return agents;
-    }
-    const hasCurrentAgent = agents.some((a) => a.name === qualifiedAgentName);
-    if (hasCurrentAgent) {
-      return agents;
-    }
-    return [
-      { name: qualifiedAgentName, headVersionId: null, updatedAt: "" },
-      ...agents,
-    ];
-  })();
+  const qualifiedAgentName = data?.agent?.name;
+  const agentOptions = agents;
 
   const handleDisconnect = () => {
     detach(
