+++ turbo/apps/platform/src/views/integrations-page/slack-settings-page.tsx
@@ -204,36 +204,8 @@ export function SlackSettingsPage() {
   const openConfirm = useSet(openSlackDisconnectDialog$);
   const closeConfirm = useSet(closeSlackDisconnectDialog$);
 
-  // Construct qualified agent name that matches the format used in agentsList$
-  // (shared agents use "org/name", owned agents use just "name").
-  const qualifiedAgentName = (() => {
-    if (!data?.agent) {
-      return undefined;
-    }
-    const fullName = `${data.agent.orgSlug}/${data.agent.name}`;
-    // If the qualified name exists in agents list, use it (shared agent)
-    if (agents.some((a) => a.name === fullName)) {
-      return fullName;
-    }
-    // Otherwise use bare name (owned agent)
-    return data.agent.name;
-  })();
-
-  // Ensure the current workspace agent appears in the dropdown even if
-  // it isn't in the user's own agents list (e.g. shared by another user).
-  const agentOptions = (() => {
-    if (!qualifiedAgentName) {
-      return agents;
-    }
-    const hasCurrentAgent = agents.some((a) => a.name === qualifiedAgentName);
-    if (hasCurrentAgent) {
-      return agents;
-    }
-    return [
-      { name: qualifiedAgentName, headVersionId: null, updatedAt: "" },
-      ...agents,
-    ];
-  })();
+  const qualifiedAgentName = data?.agent?.name;
+  const agentOptions = agents;
 
   const handleDisconnect = () => {
     detach(
